+++ scripts/archiso-base-211216.sh
@@ -24,11 +24,29 @@ rm -fr archlive
 cp -r /usr/share/archiso/configs/baseline archlive
 cd archlive
 
+# killw - kill wrapper
+killw() { pid=$1
+	echo -1:$pid $(pgrepw $pid) \
+		| xargs -n1 \
+		| sort -r \
+		| awk -F: '{print $2}' \
+		| xargs -I{} bash -c "kill -9 {} && echo '${FUNCNAME[0]}: killed: {}'"
+}
+
 # lnw - ln wrapper
 lnw() { to="$1"; from="$2"; shift; shift
 	dir=airootfs
 	mkdir -p "$(dirname "$dir/$from")"
 	ln "$@" -s "$to" "$dir/$from" && echo -e "${FUNCNAME[0]}: created "'\e[1;36msymlink\e[m'": '/$from' -> '$to'"
+}
+
+# pgrepw - pgrep wrapper
+pgrepw() { pid=$1; depth=${2:-0}
+	local d=$((depth + 1))
+	for p in $(pgrep -P $pid); do
+		echo $d:$p
+		pgrepw $p $d
+	done
 }
 
 # teew - tee wrapper
@@ -1275,7 +1293,7 @@ _
 	#Server = file:///home/custompkgs
 	/cat
 	) --dbpath $temp --noconfirm -Swy - < packages
-	kill $pid
+	killw $pid
 	rm $mirrorlist /var/cache/pacman/pkg/{community,core,extra}.db
 	repo-add live.db.tar.gz *.pkg.tar.{xz,zst}
 )